+++ api/v1alpha1/identity_provider_types.go
@@ -279,7 +279,11 @@ func (idp *IdentityProvider) ValidateCreate(ctx context.Context, obj *IdentityPr
 	allErrs = append(allErrs, result.Errors...)
 
 	// Multi-IDP: Validate Issuer field for multi-IDP mode (must be unique - requires k8s client)
-	allErrs = append(allErrs, ensureClusterWideUniqueIssuer(ctx, obj.Spec.Issuer, obj.Name, field.NewPath("spec").Child("issuer"))...)
+	issuer := obj.Spec.Issuer
+	if issuer == "" {
+		issuer = obj.Spec.OIDC.Authority
+	}
+	allErrs = append(allErrs, ensureClusterWideUniqueIssuer(ctx, issuer, obj.Name, field.NewPath("spec").Child("issuer"))...)
 
 	var warnings admission.Warnings
 	if len(allErrs) == 0 {